+++ examples/04_streamflow_intermittence_in_transient/example_04.py
@@ -24,12 +24,14 @@ import matplotlib as mpl
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
 
-import imageio
+import rasterio
 import whitebox
 wbt = whitebox.WhiteboxTools()
 wbt.verbose = False
 import xarray as xr
 xr.set_options(keep_attrs = True)
+
+import hydromodpy
 
 # ROOT DIRECTORY
 from os.path import dirname, abspath
@@ -52,7 +54,6 @@ def select_period(df, first, last):
 #%% ---- PATHS
 
 #%% PERSONAL
-
 example_path = os.path.join(root_dir, "examples", "04_streamflow_intermittence_in_transient/")
 data_path = os.path.join(example_path, "data/")
 
@@ -126,7 +127,7 @@ clim_data_mode = 'SIM2' # SIM2 online climatic data (https://meteo.data.gouv.fr/
 
 if clim_data_mode == 'SIM2':
     BV.climatic.update_sim2_reanalysis(var_list=['recharge', 'runoff',
-                                                ], 
+                                                ],
                                             nc_data_path=os.path.join(
                                                 stable_folder,
                                                 'add_data',
@@ -142,7 +143,7 @@ if clim_data_mode == 'SIM2':
     ### Units
     BV.climatic.update_recharge(BV.climatic.recharge / 1000, sim_state='transient') # from mm to m
     BV.climatic.update_runoff(BV.climatic.runoff / 1000, sim_state='transient') # from mm to m
-    
+
     ### Format for plots
     if isinstance(BV.climatic.recharge, float):
         print(f"Time-space daily average value for recharge = {BV.climatic.recharge} m")
@@ -153,22 +154,22 @@ if clim_data_mode == 'SIM2':
             r = BV.climatic.runoff.drop('spatial_ref').mean(dim = ['x', 'y']).to_pandas().iloc[:,0]
             # R = R.resample('M').sum()
             # r = r.resample('M').sum()
-        elif isinstance(BV.climatic.recharge, pd.core.series.Series):  
+        elif isinstance(BV.climatic.recharge, pd.core.series.Series):
             R = BV.climatic.recharge
             r = BV.climatic.runoff
-        
+
 elif clim_data_mode == 'local':
     x = pd.read_csv(data_path+'/'+'_climate_REANALYSIS.csv', sep=';', index_col=0)
     date_object = pd.to_datetime(x.index, format = "%d/%m/%Y")
     x.index = date_object
     x = x.sort_index()
-    x = x.resample('M').mean()
+    x = x.resample('ME').mean()
     x = select_period(x, 2000, 2002)
     BV.climatic.update_recharge(x['REC_REA_historic'] / 1000, sim_state='transient') # from mm to m
     BV.climatic.update_runoff(x['RUN_REA_historic'] / 1000, sim_state='transient') # from mm to m
     R = BV.climatic.recharge
     r = BV.climatic.runoff
-    
+
 # Plots
 fig, ax = plt.subplots(1,1, figsize=(6,3))
 ax.plot(R, label='recharge_reanalysis', c='dodgerblue', lw=2)
@@ -262,14 +263,14 @@ list_model_modflow = []
 
 for i, sy in enumerate(list_porosity[:]):
     BV.hydraulic.update_sy(sy)
-    
+
     model_name = iD_set_simulations+'_'+str(i)+'_'+str(round(sy,3))
     BV.settings.update_model_name(model_name)
     print(model_name)
-    
+
     model_modflow = BV.preprocessing_modflow(for_calib=False)
     success_modflow = BV.processing_modflow(model_modflow, write_model=True, run_model=True)
-    
+
     list_model_name.append(model_name)
     list_success_modflow.append(success_modflow)
     list_model_modflow.append(model_modflow)
@@ -306,7 +307,7 @@ for model_name, success_modflow, model_modflow in zip(list_model_name,
     if success_modflow == True:
         BV.postprocessing_modflow(model_modflow,
                                   watertable_elevation = True,
-                                  watertable_depth= True, 
+                                  watertable_depth= True,
                                   seepage_areas = True,
                                   outflow_drain = True,
                                   groundwater_flux = True,
@@ -319,10 +320,10 @@ for model_name, success_modflow, model_modflow in zip(list_model_name,
 
         timeseries_results = BV.postprocessing_timeseries(model_modflow=model_modflow,
                                                           model_modpath=None,
-                                                          datetime_format=True, 
+                                                          datetime_format=True,
                                                           subbasin_results=True,
                                                           intermittency_monthly=True) # or None
-        
+
         netcdf_results = BV.postprocessing_netcdf(model_modflow,
                                                   datetime_format=True)
 
@@ -330,8 +331,8 @@ for model_name, success_modflow, model_modflow in zip(list_model_name,
 
 #%% CROSS MIN MAX
 
-dates = pd.date_range(start='01/01/2000', end='31/12/2002', freq='M')
-    
+dates = pd.date_range(start='01/01/2000', end='31/12/2002', freq='ME')
+
 stable_folder = os.path.join(out_path, watershed_name, 'results_stable') # necessary for plots
 simulations_folder = os.path.join(out_path, watershed_name, 'results_simulations')
 
@@ -342,9 +343,9 @@ fig, axs = plt.subplots(3, 1, figsize=(5,10), dpi=300)
 axs = axs.ravel()
 
 for i, simul in enumerate(simul_list[:]):
-        
+
     model_name = os.path.split(simul)[-1]
-    
+
     ax = axs[i]
 
     Smod_path = os.path.join(simul, r'_postprocess/_timeseries/_simulated_timeseries.csv')
@@ -352,54 +353,53 @@ for i, simul in enumerate(simul_list[:]):
     Smod = Smod.reset_index()
     argmin = Smod['total_areas'].argmin()
     argmax = Smod['total_areas'].argmax()
-    
-    mask = imageio.imread(os.path.join(stable_folder, 'geographic', 'watershed_dem.tif'))
-        
-    
-    watertable_elevation = np.load(os.path.join(simulations_folder, 
+
+    with rasterio.open(os.path.join(stable_folder, 'geographic', 'watershed_dem.tif')) as src:
+        mask = src.read(1)
+
+
+    watertable_elevation = np.load(os.path.join(simulations_folder,
                                                 model_name, '_postprocess',
                                                 'watertable_elevation'+'.npy'),
                                    allow_pickle=True).item()
-    
+
     min_wt = dict()
-    
+
     cp = 0
 
     for i, key in enumerate([argmin, argmax]):
         print(key)
 
-        dem_data = imageio.imread(BV.geographic.watershed_dem)
+        with rasterio.open(BV.geographic.watershed_dem) as src:
+            dem_data = src.read(1)
         wt_data = watertable_elevation[key]
-        # river_data = imageio.imread(os.path.join(stable_folder, 'hydrography',
-        #                                           'regional stream network.tif'))
-    
         xvalues = np.linspace(-1,1,dem_data.shape[1])
         yvalues = np.linspace(-1,1,dem_data.shape[0])
         xx, yy = np.meshgrid(xvalues,yvalues)
-        
+
         cur_x = dem_data.shape[1] /2
         cur_y = dem_data.shape[0] /2 # 39
         cur_y = 39
-        
+
         dem_max = dem_data.max()
         dem_prof = dem_data.astype(float)
         dem_prof[dem_prof<0] = np.nan
         wt_prof = wt_data.astype(float)
         wt_prof[wt_prof<0] = np.nan
-        
+
         dem_h_plot = dem_prof[int(cur_y),:]
         dem_h_plot[dem_h_plot == 0] = np.nan
         wt_h_plot = wt_prof[int(cur_y),:]
         wt_h_plot[wt_h_plot == 0] = np.nan
-                    
+
         dem_max = dem_data.max()
         dem_prof = dem_data.astype(float)
         dem_prof[dem_prof<0] = np.nan
         dem_plot = np.ma.masked_array(dem_data, mask=(dem_data<0))
-        
+
         wt_prof = wt_data.astype(float)
         wt_prof[wt_prof<0] = np.nan
-        
+
         cp+=1
 
         if i == 0:
@@ -421,7 +421,7 @@ for i, simul in enumerate(simul_list[:]):
         ax.set_yticks([130,140,150,160,170])
         ax.set_xlabel('Distance [m]')
         ax.set_ylabel('Elevation [m]')
-        
+
         ax.set_title(model_name.upper(), fontsize=8)
         if i == 1:
             ax.text(0.1, 0.8, 'Max. '+str(dates[key])[:7],
@@ -430,47 +430,49 @@ for i, simul in enumerate(simul_list[:]):
             ax.text(0.1, 0.7, 'Min. '+str(dates[key])[:7],
                     transform=ax.transAxes, color='navy')
         print((str(dates[key])[:7]))
-            
+
 fig.tight_layout
-        
+
 # fig.savefig(os.path.join(simulations_folder, '_figures',
 #             'CROSS_'+iD_set_simulations+'.png'),
 #             bbox_inches='tight')
-    
+
 #%% MAP MIN MAX
 
-dates = pd.date_range(start='01/01/2000', end='31/12/2002', freq='M')
-    
+dates = pd.date_range(start='01/01/2000', end='31/12/2002', freq='ME')
+
 stable_folder = os.path.join(out_path, watershed_name, 'results_stable') # necessary for plots
 simulations_folder = os.path.join(out_path, watershed_name, 'results_simulations')
 
-line = imageio.imread(os.path.join(stable_folder, 'geographic', 'watershed_contour.tif'))
+with rasterio.open(os.path.join(stable_folder, 'geographic', 'watershed_contour.tif')) as src:
+    line = src.read(1)
 line = np.ma.masked_where(line <= 0, line)
 
-mask = imageio.imread(os.path.join(stable_folder, 'geographic', 'watershed_dem.tif'))
+with rasterio.open(os.path.join(stable_folder, 'geographic', 'watershed_dem.tif')) as src:
+    mask = src.read(1)
 
 simul_list = sorted(glob.glob(os.path.join(simulations_folder, iD_set_simulations+'*')),
                    key=os.path.getmtime)
-        
+
 for simul in simul_list[:]:
-    
+
     model_name = os.path.split(simul)[-1]
-        
+
     Smod_path = os.path.join(simul, r'_postprocess/_timeseries/_simulated_timeseries.csv')
     Smod = pd.read_csv(Smod_path, sep=';', index_col=0, parse_dates=True)
-        
+
     min_area = Smod['total_areas'].min()
     min_idx = np.argmin(Smod['total_areas'])
     max_area = Smod['total_areas'].max()
     max_idx = np.argmax(Smod['total_areas'])
     max_year = Smod['total_areas'].index[max_idx]
-    
+
     acc_npy = np.load(os.path.join(simul, '_postprocess', 'accumulation_flux.npy'), allow_pickle=True).item()
     inf = 0
     sup = 12
     compt = 0
     step = int(round(len(acc_npy)/12))
-    
+
     for i in range(step):
         print(str(i)+'/'+str(step))
         interv = list(acc_npy.items())[inf:sup]
@@ -485,31 +487,31 @@ for simul in simul_list[:]:
         days_flux = zero.copy()
         days_flux = np.ma.masked_array(days_flux, mask=(mask<0))
         days_flux = np.ma.masked_array(days_flux, mask=(days_flux<=0))
-    
+
     fig, axs = plt.subplots(1,2, figsize=(7,6))
     axs = axs = axs.ravel()
-    
+
     for k, j in enumerate([min_idx, max_idx]):
-            
+
         ax = axs[k]
-    
+
         year = Smod['total_areas'].index[j]
         val = Smod.iloc[j]['total_areas']
 
         days_flux = acc_npy[j]
-        
+
         ax.set_title(str(year)[0:10] + '   ' + '$A_{sat}$ = ' + str(val.round(1)) + ' [%]',
                      pad=10, fontsize=10)
         ax.imshow(np.ma.masked_where(mask<0, mask), cmap='Greys', alpha=0.5, zorder=0)
         ax.imshow(np.ma.masked_where((days_flux<=0) | (mask <0),
-                                     days_flux), 
+                                     days_flux),
                   cmap = mpl.colors.ListedColormap(['navy'])) # dodgerblue
         ax.imshow(line, cmap=mpl.colors.ListedColormap('k'))
-        
+
         ax.get_xaxis().set_visible(False)
         ax.get_yaxis().set_visible(False)
         ax.axis('off')
-            
+
         try:
             path_sub = os.path.join(glob.glob(
                 os.path.join(stable_folder, 'subbasin','intermittency*'))[0],
@@ -523,75 +525,79 @@ for simul in simul_list[:]:
                                        base = os.path.join(stable_folder,
                                                            'geographic',
                                                            'watershed_dem.tif'))
-            line_sub = imageio.imread(os.path.join(glob.glob(
+            with rasterio.open(os.path.join(glob.glob(
                 os.path.join(stable_folder, 'subbasin', 'intermittency*'))[0],
-                'watershed_contour.tif'))
+                'watershed_contour.tif')) as src:
+                line_sub = src.read(1)
             line_sub = np.ma.masked_where(line_sub <= 0, line_sub)
             ax.imshow(line_sub, cmap=mpl.colors.ListedColormap('grey'))
         except:
             pass
-        
+
     fig.suptitle(model_name.upper(), y=0.85, fontsize=8)
-    
+
     fig.tight_layout()
-                
+
     # fig.savefig(os.path.join(simulations_folder, '_figures',
     #             'MAPminmax_'+model_name+'.png'),
     #             bbox_inches='tight')
 
 #%% PERSISTENCY
 
-simul_list = sorted(glob.glob(os.path.join(simulations_folder, 
+simul_list = sorted(glob.glob(os.path.join(simulations_folder,
                                            iD_set_simulations+'*')),
                    key=os.path.getmtime)
 
-line = imageio.imread(os.path.join(stable_folder,
-                                   'geographic',
-                                   'watershed_contour.tif'))
+with rasterio.open(os.path.join(stable_folder,
+                                'geographic',
+                                'watershed_contour.tif')) as src:
+    line = src.read(1)
 line = np.ma.masked_where(line <= 0, line)
 
-mask = imageio.imread(os.path.join(stable_folder,
-                                   'geographic',
-                                   'watershed_dem.tif'))
+with rasterio.open(os.path.join(stable_folder,
+                                'geographic',
+                                'watershed_dem.tif')) as src:
+    mask = src.read(1)
 
 fig, axs = plt.subplots(1, 3, figsize=(7,6))
 axs = axs = axs.ravel()
 
 for i, simul in enumerate(simul_list[:]):
-        
+
     model_name = os.path.split(simul)[-1]
 
     ax = axs[i]
-        
-    pi = imageio.imread(os.path.join(simul, r'_postprocess/_rasters',
-                                     'persistency_index_t(-).tif'))
+
+    with rasterio.open(os.path.join(simul, r'_postprocess/_rasters',
+                                    'persistency_index_t(-).tif')) as src:
+        pi = src.read(1)
     pi = np.ma.masked_where(pi==-9999, pi)
     pi = np.ma.masked_where(mask==-99999, pi)
-    
+
     im = ax.imshow(pi, cmap='jet')
-    
+
     ax.imshow(line, mpl.colors.ListedColormap(['k']),
                   vmin=0, vmax=1)
-    
+
     ax.get_xaxis().set_visible(False)
     ax.get_yaxis().set_visible(False)
     ax.axis('off')
-    
+
     ax.set_title(model_name.upper(), fontsize=8)
-    
+
     # fig.subplots_adjust(right=0.8)
     cbar_ax = fig.add_axes([0.25, 0.25, 0.5, 0.02])
     cb = fig.colorbar(im, cax=cbar_ax, orientation="horizontal", pad=0.2)
     cb.set_label('Persistency index [-]', fontsize=10)  # cax == cb.ax
-        
+
     fig.tight_layout()
 
 fig.tight_layout
-        
+
 # fig.savefig(os.path.join(simulations_folder, '_figures',
 #             'PI'+iD_set_simulations+'.png'),
 #             bbox_inches='tight')
-    
+
 #%% STREAMFLOW
 
 # iD_set_simulations = 'explorSy_mperday_monthly_steady'
@@ -604,29 +610,29 @@ Qobs = pd.read_csv(Qobs_path, sep=';', index_col=0, parse_dates=True)
 
 area = int(round(BV.geographic.area))
 Qobs = (Qobs / (area*1000000)) * (3600 * 24) # m3/s to m/day
-Qobs = Qobs.resample('M').sum() * 1000 # m/day to mm/month
+Qobs = Qobs.resample('ME').sum() * 1000 # m/day to mm/month
 
 simul_list = sorted(glob.glob(os.path.join(simulations_folder,
                                            iD_set_simulations+'*')),
                    key=os.path.getmtime)
 
 for i, simul in enumerate(simul_list[:]):
-    
+
     fig, (a0, a1) = plt.subplots(1, 2, gridspec_kw={'width_ratios': [3, 1]},
                                  figsize=(10,3))
 
     model_name = os.path.split(simul)[-1]
-        
-    Smod_path = os.path.join(simul, 
+
+    Smod_path = os.path.join(simul,
                              r'_postprocess/_timeseries/_simulated_timeseries.csv')
     Smod = pd.read_csv(Smod_path, sep=';', index_col=0, parse_dates=True)
-    
-    Qmod = Smod['outflow_drain'] 
+
+    Qmod = Smod['outflow_drain']
     Qmod = Qmod.squeeze() * 1000
     Qmod = (Qmod + (r * 1000)) * Qmod.index.day
-    
+
     Rmod = Smod['recharge'] * Qmod.index.day
-    
+
     yearsmaj = mdates.YearLocator(1)   # every year
     yearsmin = mdates.YearLocator(1)
     # monthsmaj = mdates.MonthLocator(6)  # every month
@@ -648,26 +654,27 @@ for i, simul in enumerate(simul_list[:]):
     ax.set_xlim(pd.to_datetime('1999'), pd.to_datetime('2004'))
     ax.legend()
     ax.set_title(model_name.upper(), fontsize=10)
-    
+
     axb = ax.twinx()
     axb.bar(Rmod.index, Rmod,color='blue', edgecolor='blue', lw=2.5)
     axb.set_ylim(0,999)
     axb.invert_yaxis()
-    axb.set_yticklabels([0,200])
-    
+    axb.set_yticks([0, 200])
+    axb.set_yticklabels([0, 200])
+
     Qobs_stat = select_period(Qobs, 2000, 2002)
     Qmod_stat = select_period(Qmod, 2000, 2002)
-    
+
     Qmod_stat = pd.DataFrame(Qmod_stat)
     Qmod_stat = Qmod_stat.set_index(Qobs_stat.index)
-    
+
     rmse_val, nrmse, nse_val, nselog_val, bal, mare_val, kge_val = toolbox.efficiency_criteria(Qmod_stat, Qobs_stat)
     print(model_name.upper())
     print(round(nse_val,2))
     print(round(nselog_val,2))
     print(round(rmse_val,2))
     print(round(kge_val,2))
-    
+
     ax = a1
     ax.scatter(Qobs_stat, Qmod_stat,
                s=25, edgecolor='none', alpha=0.75, facecolor='forestgreen')
@@ -677,12 +684,12 @@ for i, simul in enumerate(simul_list[:]):
     ax.set_xlim(1,500)
     ax.set_ylim(1,500)
     # ax.set_xlim(0.1,300)
-    # ax.set_ylim(0.1,300)    
+    # ax.set_ylim(0.1,300)
     ax.set_xlabel('$Q_{obs}$ / A [mm/month]', fontsize=12)
     ax.set_ylabel('$Q_{sim}$ / A [mm/month]', fontsize=12)
 
     fig.tight_layout()
-                
+
     # fig.savefig(os.path.join(simulations_folder, '_figures',
     #             'STREAMFLOW_'+model_name+'.png'),
     #             bbox_inches='tight')
@@ -698,26 +705,26 @@ Qobs = pd.read_csv(Qobs_path, sep=';', index_col=0, parse_dates=True)
 
 area = int(round(BV.geographic.area))
 Qobs = (Qobs / (area*1000000)) * (3600 * 24) # m3/s to m/day
-Qobs = Qobs.resample('M').sum() * 1000 # m/day to mm/month
+Qobs = Qobs.resample('ME').sum() * 1000 # m/day to mm/month
 
 simul_list = sorted(glob.glob(os.path.join(simulations_folder,
                                            iD_set_simulations+'*')),
                    key=os.path.getmtime)
 
 for i, simul in enumerate(simul_list[:]):
-    
+
     model_name = os.path.split(simul)[-1]
-    
-    Smod_path = os.path.join(simul, 
+
+    Smod_path = os.path.join(simul,
                              r'_postprocess/_timeseries/_simulated_timeseries.csv')
     Smod = pd.read_csv(Smod_path, sep=';', index_col=0, parse_dates=True)
-    
-    Qmod = Smod['outflow_drain'] 
+
+    Qmod = Smod['outflow_drain']
     Qmod = Qmod.squeeze() * 1000
     Qmod = Qmod + (r * 1000)
-    
+
     Rmod = Smod['recharge'] * 1000
-        
+
     Sonde_path = os.path.join(glob.glob(
         os.path.join(simul, r'_subbasins/intermittency_*'))[0],
         '_simulated_timeseries.csv')
@@ -732,16 +739,16 @@ for i, simul in enumerate(simul_list[:]):
     # accep = d[d==4].dropna()
     # visib = d[d==5].dropna()
     # d = d.resample('M').mean()
-    
+
     # Smod['onde'] = d
-    
+
     from datetime import timedelta
     x_months = Smod.index + timedelta(days=-30)
     Smod['date'] = x_months
     Smod.index = Smod['date']
-    
+
     fig, ax = plt.subplots(1, 1, figsize=(7,3.5))
-    
+
     ax.fill_between(Smod.index, 0, Smod['total_areas'],
                     interpolate=False, color='dodgerblue', alpha=0.5,
                     step='pre', label='Intermittent part')
@@ -768,15 +775,15 @@ for i, simul in enumerate(simul_list[:]):
     months_maj = mdates.MonthLocator()  # every x month
     ax.xaxis.set_major_locator(years_maj)
     ax.xaxis.set_minor_locator(months_maj)
-    
+
     ax.set_title(model_name.upper(), fontsize=10)
-    
+
     fig.tight_layout()
-                
+
     # fig.savefig(os.path.join(simulations_folder, '_figures',
     #             'SATURATION_'+model_name+'.png'),
     #             bbox_inches='tight')
-                
+
 #%% ---- NOTES
 
 os.chdir(root_dir)